fixup! some fixes

+++ systemtest/src/test/java/io/strimzi/systemtest/kafka/KafkaST.java
@@ -1665,7 +1665,9 @@ class KafkaST extends AbstractST {
             cluster.deleteCustomResources(TEMPLATE_PATH);
         }
 
-        cmdKubeClient().deleteByName(Kafka.RESOURCE_KIND, OPENSHIFT_CLUSTER_NAME);
+        if (KafkaResource.kafkaClient().inNamespace(NAMESPACE).withName(OPENSHIFT_CLUSTER_NAME).get() != null) {
+            cmdKubeClient().deleteByName(Kafka.RESOURCE_KIND, OPENSHIFT_CLUSTER_NAME);
+        }
 
         kubeClient().listPods().stream()
             .filter(p -> p.getMetadata().getName().startsWith(OPENSHIFT_CLUSTER_NAME))
